Added get all customers button
Implemented a button to the interface that allows a get all request from the API. Also added a textbox in order to display the returned information in. The test API connectivity button will now display the get all button as well as the text box and hide itself once clicked.