Implemented update customer by name
Added appropriate labels, text boxes, and buttons in order to allow the user to update a customer by name. When clicking update by name the controls appear then after clicking the update button if it is successful it will send a message saying updated and will hide all of the controls again. Also implemented some methods for showing and hiding controls.